feat: add external celery workers to the processing services
This reverts commit 460f27c537b829f05416cf736cf50185cc7b136b.
