+++ processing_services/example/Dockerfile
@@ -1,7 +1,11 @@
 FROM python:3.11-slim
 
-# Set up ml backend FastAPI
 WORKDIR /app
+
 COPY . /app
+
 RUN pip install -r ./requirements.txt
+
+RUN chmod +x ./celery_worker/start_celery.sh
+
 CMD ["python", "/app/main.py"]
