+++ processing_services/minimal/Dockerfile
@@ -6,4 +6,6 @@ COPY . /app
 
 RUN pip install -r ./requirements.txt
 
+RUN chmod +x ./celery_worker/start_celery.sh
+
 CMD ["python", "/app/main.py"]
