+++ processing_services/minimal/requirements.txt
@@ -3,3 +3,5 @@ uvicorn==0.35.0
 pydantic==2.11.7
 Pillow==11.3.0
 requests==2.32.4
+celery==5.4.0
+redis==5.2.1
